+++ controllers/chat.js
@@ -1,7 +1,7 @@
-const TwilioUtil = require('../util/twilio');
+const TwilioService = require('../services/twilio');
 
 exports.createToken = function (req, res) {
-    const token = TwilioUtil.TokenGenerator(req.user._id, req.body.device);
+    const token = TwilioService.TokenGenerator(req.user._id, req.body.device);
 
     res.json({
         _id: req.user._id,
